requirements.txt
openai==1.3.0
PyYAML==6.0.1
gradio==4.16.0
langchain==0.0.350
